+++ mobile/android/android-components/components/concept/engine/src/main/java/mozilla/components/concept/engine/history/HistoryTrackingDelegate.kt
@@ -4,7 +4,7 @@
 
 package mozilla.components.concept.engine.history
 
-import mozilla.components.concept.storage.VisitType
+import mozilla.components.concept.storage.PageVisit
 
 /**
  * An interface used for providing history information to an engine (e.g. for link highlighting),
@@ -17,7 +17,7 @@ interface HistoryTrackingDelegate {
     /**
      * A URI visit happened that an engine considers worthy of being recorded in browser's history.
      */
-    suspend fun onVisited(uri: String, type: VisitType)
+    suspend fun onVisited(uri: String, visit: PageVisit)
 
     /**
      * Title changed for a given URI.